Refactor Link Token List components architecture
