+++ client/src/pages/DeleteTokenPage/DeleteTokenItem.tsx
@@ -1,11 +1,12 @@
 import React from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import classnames from 'classnames';
-import TokenBadge from '../../Badges/TokenBadge/TokenBadge';
+
+import TokenBadge from '../../components/Badges/TokenBadge/TokenBadge';
 
 import { CheckedTokenType } from './types';
 
-import classes from './DeleteToken.module.scss';
+import classes from './styles/DeleteToken.module.scss';
 
 type Props = {
   hasLowBalance: boolean;
