+++ client/src/pages/DeleteTokenPage/styles/DeleteToken.module.scss
