--- client/src/pages/LinkTokenPages/AddTokenPage/index.ts
@@ -1,3 +1,0 @@
-import AddTokenPage from './AddTokenPage';
-
-export default AddTokenPage;
