--- client/src/pages/LinkTokenPages/DeleteTokenPage/index.ts
@@ -1,3 +1,0 @@
-import DeleteTokenPage from './DeleteTokenPage';
-
-export default DeleteTokenPage;
