--- client/src/pages/LinkTokenPages/TokenListPage/index.ts
@@ -1,3 +1,0 @@
-import TokenListPage from './TokenListPage';
-
-export default TokenListPage;
