+++ client/src/routes.jsx
@@ -19,9 +19,9 @@ import FioDomainTransferPage from './pages/FioDomainTransferPage';
 import FioDomainStatusChangePage from './pages/FioDomainStatusChangePage';
 import FioAddressRenewPage from './pages/FioAddressRenewPage';
 import FioDomainRenewPage from './pages/FioDomainRenewPage';
-import TokenListPage from './pages/LinkTokenPages/TokenListPage';
-import DeleteTokenPage from './pages/LinkTokenPages/DeleteTokenPage';
-import AddTokenPage from './pages/LinkTokenPages/AddTokenPage';
+import TokenListPage from './pages/LinkTokenPage';
+import DeleteTokenPage from './pages/DeleteTokenPage';
+import AddTokenPage from './pages/AddTokenPage';
 import SettingsPage from './pages/SettingsPage';
 import RefHomePage from './pages/RefHomePage';
 import RefSignNftPage from './pages/RefSignNftPage';
